Fix profile picture not displayed
